rag/pyproject.toml
[project]
name = "scorebase-rag"
version = "0.1.0"
description = "RAG-powered semantic search for ScoreBase"
requires-python = ">=3.11"
dependencies = [
    "haystack-ai>=2.8.0",
    "chroma-haystack>=1.0.0",
    "music21>=9.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
